Make noAnimations imply noink on the inner paper-ripple.
Fixes #102 where animation-free dropdowns ended up
with tons of Ripple objects endlessly and invisibly animating in
requestAnimationFrames, though I'll admit it's not a principled fix,
in that I'm not sure what the root cause of the bug is.

This seems to be roughly in line with user expectations I think, as
someone who doesn't want animations probably also doesn't want a
ripple animation.

I went this route rather than making `noink` work because the noink
property isn't mapped anywhere, and the requestAnimationFrame leak
would still occur for people using noAnimations.