Fix accessibility descriptions of paper-dropdown-menu-light. (...)
paper-dropdown-menu-light is not a combobox: a combobox contains a
user-editable text area for arbitrary inputs with a popup for suggestions
(usually a 'listbox') and this element does not allow arbitrary user input.
paper-dropdown-menu-light is a 'collapsible dropdown listbox'. See this link
for an example of this behavior:
https://www.w3.org/TR/wai-aria-practices-1.1/examples/listbox/listbox-collapsible.html.

Given these issues with the current state of the element, I've made these
changes:

- Update role from "combobox" to "button".

  This element isn't a combobox: it is a button that opens a popup of some
  kind, typically containing a listbox.

- Remove the `aria-autocomplete` attribute.

  There is nothing to autocomplete because this element does not contain a
  user-editable text input.

- Set the contained paper-menu-button's role to "none".

  Otherwise, paper-menu-button sets its own role to "group" and prevents
  VoiceOver from reading its content as the label of the
  paper-menu-dropdown-light (acting as a button). Also, the paper-menu-button
  itself isn't focusable, so when a user arrives at the
  paper-dropdown-menu-light in the sequential focus order, _it_ becomes focused
  rather than the paper-menu-button.

- Remove an empty span with `role="button"` inside the shadow root.

  There's an explicit warning about not removing it but I don't have any reason
  to believe it's actually useful: the warning explicitly calls out that it's
  fulfilling an accessibility requirement for a _combobox_ - which this element
  is not - and the span isn't focusable.

Also, even though this element and its demos imply that it acts as a
listbox-popup-opening button, the current element only sets `aria-haspopup` to
`true` rather than explicitly "listbox". So, I've decided not to change it in
case someone using this element is relying on the current generic value of
`true` and showing something other than a "listbox".