Move the focusable from the host to the dropdown trigger. (...)
In the non-light paper-dropdown-menu, all of the styling is handled by the
wrapped paper-input, which assumes that its internal input is what will act as
the focusable element for the control. If an input is focusable but you want
the input to act as something other than whatever is specified by its `type`
attribute, then you need to explicitly set the `role` attribute on that input
as well as any other `aria-` attributes relevant to that particular role.

Also, users seem to take advantage of the fact that our non-standard CSS mixins
(`@apply --mixin;`) inherit and often style the paper-dropdown-menu by setting
mixins intended to inherit to the paper-input. So, removing the paper-input
completely (e.g. trying to do something like replacing paper-dropdown-menu with
paper-dropdown-menu-light) would likely break a lot of users.

Given all of these constraints, the only reasonable way I've found so far to
(a) keep the paper-input (for its styles), (b) keep the paper-input's wrapped
input as the focusable element (to prevent paper-input's styling behavior from
breaking), and (c) fix the incorrect accessibility descriptions is to update
paper-input to accept a new set of properties to pass values for the `role` and
all associated `aria-` attributes so they can be set on the internal input.

Fortunately, for the sake of making paper-input act as a menu button, it only
needs to pass through the `role` and `aria-haspopup` attributes. However, these
aren't currently in paper-input so this commit won't work until I've gotten
those properties added to paper-input.

Also, I still need to figure out if it's too dangerous to move the focusable
element from the host to an internal element because users applying different
`role` and `aria-` attributes to the host won't be applying them to the correct
element anymore.